generaDBScript/requirements.txt
click==8.1.7
PyYAML==6.0.2
pydantic==2.8.2
pyodbc==4.0.39
Jinja2==3.1.4
pytest==8.2.2
